pyproject.toml
[tool.poetry]
name = "fastapi-practice"
version = "0.1.0"
description = ""
authors = ["Stanislav Khlud <stanislav.khlud@saritasa.com>"]

[tool.poetry.dependencies]
python = "^3.10"
rich = "^11.0.0"
fastapi = "^0.71.0"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
arrow = "^1.2.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
